Create Image_Classification
The goal of the project is to develop a deep learning model that can accurately classify images into predefined categories or classes. This involves several steps, starting with data collection and preprocessing, where a diverse and representative dataset of images is gathered and prepared for training. The next step is model architecture design, where a deep neural network structure is defined, typically consisting of multiple layers such as convolutional layers, pooling layers, and fully connected layers